Luneetee3D/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.22.0
pillow>=9.0.0
trimesh>=3.9.0
matplotlib>=3.5.0
tqdm>=4.62.0
pyyaml>=6.0
opencv-python>=4.5.0
scipy>=1.7.0
diffusers>=0.14.0
transformers>=4.25.0
einops>=0.4.0
omegaconf>=2.1.0
pymeshlab>=2022.0
pygltflib>=1.15.0
xatlas>=0.0.6
accelerate>=0.16.0
rembg>=2.0.0
scikit-learn>=1.0.0
scikit-image>=0.19.0
ninja>=1.10.0
pybind11>=2.9.0
